Move middleware code to common/server
