--- backend/internal/common/auth/doc.go
@@ -1,2 +1,0 @@
-// Package auth provides shared authentication session primitives and API middleware.
-package auth
