+++ backend/internal/common/server/auth/doc.go
@@ -1,0 +1,2 @@
+// Package auth provides shared server authentication session primitives and API middleware.
+package auth
