+++ backend/internal/common/server/middleware/doc.go
@@ -1,0 +1,2 @@
+// Package middleware provides shared HTTP middleware.
+package middleware
