--- backend/internal/server/middleware/doc.go
@@ -1,3 +1,0 @@
-// Package middleware provides HTTP middleware used by the server package.
-// E.g., a logging for each request.
-package middleware
